Edição de Anexos (imagem e descricao) COMPLETO, possibilidade de inserção de anexo sem arquivo COMPLETO e ajuste de Dashboard para o cliente COMPLETO

+++ model/ws/processaArquivos.php
@@ -8,6 +8,9 @@ include('../../config/conexao.php');
 $conexao = new classeConexao();
 
 $projetos = $conexao::exec("DELETE FROM tb_arquivos WHERE id={$id}");
-unlink('..\..\view\images\uploads\anexos/'.$nome);
+
+if($nome!='sem-anexo.jpg') { //só exclui se não for a imagem de sem anexo
+    unlink('..\..\view\images\uploads\anexos/'.$nome);
+}
 
 echo json_encode(array("status" => 'true'));
